requirements.txt
asgiref==3.7.2
Django==4.1.2
django-ckeditor==6.7.1
django-jazzmin==2.6.1
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
uritemplate==4.1.1
